+++ airone/urls.py
@@ -19,6 +19,7 @@ urlpatterns = [
     url(r'^api/v1/', include(api_v1_urlpatterns)),
     url(r'^job/', include(('job.urls', 'job'))),
     url(r'^auth/', include(('django.contrib.auth.urls', 'auth'))),
+    url(r'^webhook/', include(('webhook.urls', 'entity'))),
 ]
 
 for extension in settings.AIRONE['EXTENSIONS']: